+++ src/index.html
@@ -4,18 +4,24 @@
     <meta charset="UTF-8" />
     <meta name="viewport" content="width=device-width, initial-scale=1.0" />
     <title>Filmoteka</title>
-    <link href="https://fonts.googleapis.com/css2?family=Roboto:wght@400;500;700&display=swap" rel="stylesheet">
+    <link
+      href="https://fonts.googleapis.com/css2?family=Roboto:wght@400;500;700&display=swap"
+      rel="stylesheet"
+    />
     <link
       rel="stylesheet"
       href="https://cdnjs.cloudflare.com/ajax/libs/modern-normalize/1.0.0/modern-normalize.min.css"
       integrity="sha512-ISS7cAi1PEhQ8jnbJpJZMd29NlhNj4AWYyLOSp2CE/CsHxTCu+r+t0D2yoJudVrd0/8fTVPUVDzY5Tvli75u/g=="
-      crossorigin="anonymous"/>
-      <link rel="stylesheet" href="./sass/index.scss" />
+      crossorigin="anonymous"
+    />
+    <link rel="stylesheet" href="./sass/index.scss" />
   </head>
   <body>
     <include src="./partials/header.html"></include>
     <main>
       <section class="container">
+        <!-- my temp -->
+        <include src="./partials/temp_auth.html"></include>
         <div class="gallery" id="home-gallery"></div>
       </section>
       <include src="./partials/main.html"></include>
@@ -24,4 +30,3 @@
     <script src="./index.js" type="module"></script>
   </body>
 </html>
-
