+++ src/partials/temp_auth.html
@@ -1,0 +1,6 @@
+<div class="auth">
+  <div class="container"></div>
+  <button class="login">Login</button>
+  <button class="logout">Logout</button>
+  <button class="add">Add something</button>
+</div>
